refactor: 프로젝트 카드 클릭 훅 상위 폴더로 이동 (#174)
